+++ ui/ut-bulk/ui/react/containers/Batch/Grid/actions.js
@@ -9,7 +9,6 @@ export function fetchBatches (params) {
 }
 
 export function updatePagination (params) {
-  debugger
   return {
     type: actionTypes.BATCH_UPDATE_PAGINATION,
     method: 'bulk.batch.fetch',