+++ ui/ut-notifications/ui/react/containers/Reports/Pagination/index.js
@@ -3,27 +3,26 @@ import {connect} from 'react-redux'
 import Pagination from '../../../components/Pagination'
 import {updatePagination as update} from '../Grid/actions'
 
-class BulkPaymentPagination extends Component {
+class NotificationsPagination extends Component {
   render () {
     return (
       <div>
-        <Pagination pagination={this.props.pagination.concat(this.props.params)} onUpdate={this.props.update} />
+        <Pagination pagination={this.props.pagination} onUpdate={this.props.update} />
       </div>
     )
   }
 }
 
-BulkPaymentPagination.propTypes = {
+NotificationsPagination.propTypes = {
   pagination: PropTypes.object,
-  update: PropTypes.func,
-  params: PropTypes.object
+  update: PropTypes.func
 }
 
 export default connect(
   (state, ownProps) => {
     return {
-      pagination: state.bulkPaymentGrid.get('pagination')
+      pagination: state.notificationsReportsGrid.get('pagination')
     }
   },
   {update}
-)(BulkPaymentPagination)
+)(NotificationsPagination)